Remove and rename file
